+++ files/scripts/setup-repos-and-rpms.sh
@@ -140,7 +140,9 @@ install_available \
 echo "--- Installing desktop integration ---"
 
 install_available \
-  mpv kdeconnectd libnotify systemd-ukify
+  mpv kdeconnectd libnotify systemd-ukify \
+  nautilus gvfs gvfs-mtp gvfs-smb gvfs-afp gvfs-archive gvfs-fuse \
+  gvfs-nfs gvfs-goa gvfs-gphoto2 gnome-disk-utility dolphin
 
 echo "--- Installing developer utilities ---"