+++ openqa/main.pm
@@ -1,7 +1,0 @@
-use Mojo::Base -strict;
-use testapi;
-use autotest;
-
-autotest::loadtest 'tests/boot.pm';
-
-1;